[REMANIEMENT] Isole le code API du téléversement de services
… pour faire de la place pour le code du téléversement de modèles de mesure qui arrive.
